修改投掷基础伤害，使创造模式左键不进行操作，使余烬重锤耐久过低时会正确失效
